feat: move modules to modules folder
