+++ server/src/modules/core/routes.ts
