+++ server/src/modules/sample/v1/controllers/sample.ts
