+++ server/src/modules/sample/v1/sample.types.ts
