+++ server/tsconfig.json
@@ -30,15 +30,13 @@
 		],
 		"paths": {
 			"@app": ["src/app"],
+			"@modules": ["src/modules"],
+			"@modules/*": ["src/modules/*"],
 			"@config": ["src/config"],
 			"@config/*": ["src/config/*"],
-			"@core": ["src/core"],
-			"@core/*": ["src/core/*"],
 			"@mocks": ["test/mocks"],
 			"@mocks/*": ["test/mocks/*"],
 			"@pkg": ["package.json"],
-			"@sample": ["src/sample"],
-			"@sample/*": ["src/sample/*"],
 			"@shared": ["src/shared"],
 			"@shared/*": ["src/shared/*"],
 			"@test": ["test"],
